requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-dotenv==1.0.1
gunicorn==23.0.0
drf-spectacular==0.28.0
